Added A* Pathfinding Package, animated Water Snake
So we now have a package that can handle pathfinding for us (yay, no AI programming (yet)!) and I figured out how to apply an animated sprite to an object directly from Aseprite. To check these out go to the Jonatan/Jonathan test scene. The "Funny Button" spawns in other Water Snakes but I still haven't added the ability for them to insta-lock onto the player, so it's just there to make the framework for interactions.