Made interaction UI pop up when coming into contact with an interactable object
A UI text box will now appear when getting close to an object with a UIInteractable script. In this case the FunnyButton's Button Script is programmed to be built off the UIInteractable interface, so the Player's Interactor script recognizes it as a UIInteractable script and takes the text prompt programmed into the button and displays it on the prompt.